% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_means.R
\name{compare_means}
\alias{compare_means}
\title{Function: compare_means}
\usage{
compare_means(stats1, stats2, alpha = 0.05, alternative = "two.sided")
}
\arguments{
\item{stats1}{An object of class 'StatsSummary'.}

\item{stats2}{An object of class 'StatsSummary'.}

\item{alpha}{Significance level for the t-test (default is 0.05).}

\item{alternative}{Specifies the alternative hypothesis: "two.sided" (default), "greater", or "less".}
}
\value{
A list containing mean difference, p-value, and test conclusion.
}
\description{
Function `compare_means` compares the means of two distributions by
performing a t-test on the data stored in objects of class 'StatsSummary'.
This function checks if the difference in means of two samples is statistically
significant. That means that test checks if mean_1 = mean_2 approximately.
}
\examples{
stats1 <- calculate_stats(rnorm(50))
stats2 <- calculate_stats(rnorm(50, mean = 0.5))
result <- compare_means(stats1, stats2)
print(result)
}
