% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Dataset}
\format{
A data frame with 100 observations and 3 variables:
\describe{
  \item{x}{Randomly generated x-values from a normal distribution N(50, 10).}
  \item{y}{Randomly generated y-values from a normal distribution N(100, 20).}
  \item{z}{Randomly generated z-values from a uniform distribution U(0,1).}
}
}
\usage{
data (example_data)
}
\description{
A dataset containing 100 random points from normal and uniform distributions.
}
\examples{
data(example_data)
plot_data(example_data, "x", "y")
}
\keyword{dataset}
\keyword{example_data}
