% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_data}
\alias{plot_data}
\title{Function: plot_scatter}
\usage{
plot_data(data, x_column = "x", y_column = "y")
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{x_column}{The name of the column to be used for the x-axis. (default is "x")}

\item{y_column}{The name of the column to be used for the y-axis. (default is "y")}
}
\value{
An interactive plotly scatter plot.
}
\description{
This function creates an interactive scatter plot using the specified columns from the data.
The columns are identified by their names passed as arguments to the function.
}
\examples{
data <- data.frame(x = rnorm(100), y = rnorm(100))
plot_data(data, x_column = "x", y_column = "y")
}
